Update run-agent-locally.md
Rewording the defender text. It makes it sound like you won't get the error message if you just verify the hash and that's not true. At the moment you'll still get the defender error if you validate the hash. This is just the method we recommend for validating you're running the code you expect to be running for our service. Additionally we're working on signing it and will update the docs about that when that's compete.